Move target tolerance to cairo_test_target structure (should let single-pixel SVG errors pass)
Previously we were setting the target tolerance based on the surface
type. But that doesn't work as multiple backends will provide a surface
of type meta. So instead we put the tolerance as a value in the
cairo_test_target data structure.

With this change, some single-pixel errors of 1 in the SVG backend
should now be ignored.

Manually "cherry picked" from bcb7863f00b4cfdf0985993067fc32d07b81540b
(the regular cherry-pick command cannot handle the file renames and
 indentation changes involved here).