do not ignore errors in glyph paths
When interpreting glyph paths in ATSUI callbacks we were ignoring
the cairo status. We need to return this to the caller. In order
to do this we introduce a custom OSStatus code in the range that
Apple reccommend.
(cherry picked from commit b498c928be73635614596243d17a922e9d5ff008)